Implemented EGL_NV_post_sub_buffer.
The change to the sample is just to demonstrate how I tested it. I won't check in the stuff under samples/.

It's not clear to me what eglPostSubBufferNV should do if EGL_POST_SUB_BUFFER_SUPPORTED_NV is false. At the moment it fails silently as though it was called on the wrong surface type:

 "If <surface> is a back-buffered surface, then the requested portion
    of the color buffer is copied to the native window associated with
    that surface. If <surface> is a single-buffered window, pixmap, or
    pbuffer surface, eglSwapBuffers and eglPostSubBufferNV have no
    effect."
Review URL: http://codereview.appspot.com/5345050