Fixed leak of render target associated with ColorBuffer.
Following r853, I am assuming that any render target returned by FrameBuffer::getRenderTarget has to be released. Updating the usage of getRenderTarget to reflect that.

This is not reproducible in Chrome at the moment but it will be once I start rendering to pbuffers.

I ran the WebGL conformance tests. No regressions or crashes.
Review URL: http://codereview.appspot.com/5451110