In generated shaders, output +INF and -INF as largest single precision floating point number.
C++ streams seem to use the representation 1.$ for INF and that isn't valid syntax in GLSL or HLSL.

Also preserve the sign of INF in constant expressions that divide by zero. I can't figure out what to do about 0/0 because the shader models we are using do not support NaN. Treating it as +INF as before.
Review URL: https://codereview.appspot.com/7057046